feat(AutoScroll): Move action/mutation/state from Media to Ui and Update i18n text
