+++ components/ui/ChatScroll/ChatScroll.html
@@ -1,13 +1,13 @@
 <div :class="classObject" ref="scrollRef" v-on:scroll="onScrolled">
   <slot />
   <div
-    v-if="$store.state.media.unreadMessage>0"
+    v-if="$store.state.ui.unreadMessage>0"
     class="new-message-alert"
     :style="{ right: newMessageAlertPos.right + 'px', 
     top: newMessageAlertPos.top + 'px', width: newMessageAlertPos.width + 'px' }"
     @click="autoScrollToBottom"
   >
-   {{$store.state.media.unreadMessage}} {{ $t('pages.chat.new_message') }}
-    <span>{{ $t('pages.chat.mark_as_read') }}</span>
+    {{ $store.state.ui.unreadMessage }} {{ $t('pages.chat.new_message') }}
+    <span>{{ $t('pages.chat.jump_to_current') }}</span>
   </div>
 </div>
