+++ pages/server/chat/Chat.html
@@ -5,5 +5,5 @@
     &nbsp;
     {{$t('conversation.encrypted')}}
   </div>
-  <TailoredMessagingConversation :messages="$store.state.media.messages" :loading="loading" />
+  <TailoredMessagingConversation :messages="$store.state.ui.messages" :loading="loading" />
 </div>
