+++ store/media/actions.ts
@@ -1,3 +1,4 @@
+// eslint-disable-next-line import/named
 import { Commit } from 'vuex'
 
 interface FetchCallsArguments {
@@ -7,19 +8,10 @@ interface FetchCallsArguments {
 
 export default {
   handler: () => {},
-  async acceptCall({ commit }: FetchCallsArguments) {
+  acceptCall({ commit }: FetchCallsArguments) {
     commit('toggleIncomingCall', '')
   },
-  async denyCall({ commit }: FetchCallsArguments) {
+  denyCall({ commit }: FetchCallsArguments) {
     commit('toggleIncomingCall', '')
-  },
-  setMessages({ commit }: FetchCallsArguments, messages: any[]) {
-    commit('setMessages', messages)
-  },
-  sendMessage({ commit }: FetchCallsArguments, message: any) {
-    commit('sendMessage', message)
-  },
-  setIsScrollOver({ commit }: FetchCallsArguments, status: boolean) {
-    commit('setIsScrollOver', status)
   },
 }
